requirements copy.txt
requests
fastapi
uvicorn
mysql-connector-python
openapi-client
flask
ctrader-open-api
service_identity
sqlalchemy
aiomysql
greenlet
python-dotenv